+++ src/Layer/Layer.h
@@ -11,10 +11,10 @@ namespace jnetwork
 class Layer
 {
   public:
-    explicit Layer(const uint32_t size);
+    explicit Layer(const size_t size);
 
   private:
-    uint32_t m_Size;
+    size_t m_Size;
     std::vector<std::shared_ptr<Neuron>> m_Neurons;
 };
 } // namespace jnetwork
